+++ src/main.cpp
@@ -13,7 +13,8 @@ int main(int argc, char ** argv)
         << "Or use --list-commands to print the list of commands and their info." << endl;
         return 0;
     }
-    
+
+    string command_map_path = get_cmd_map_abs_path(&argc, argv);
     string lib_name = get_device_lib_name(&argc, argv);
     bool bypass_range_check = get_bypass_range_check(&argc, argv);
 
@@ -36,7 +37,7 @@ int main(int argc, char ** argv)
         }
     }
 
-    dl_handle_t cmd_map_handle = load_command_map_dll();
+    dl_handle_t cmd_map_handle = load_command_map_dll(command_map_path);
 
     if(next_cmd[0] == '-')
     {
@@ -47,8 +48,8 @@ int main(int argc, char ** argv)
             return print_command_list();
         }
     }
-
-    dl_handle_t device_handle = get_dynamic_lib(lib_name);
+    string lib_path = get_dynamic_lib_path(lib_name);
+    dl_handle_t device_handle = get_dynamic_lib(lib_path);
     print_args_fptr print_args = get_print_args_fptr(cmd_map_handle);
     check_range_fptr check_range = get_check_range_fptr(cmd_map_handle);
     device_fptr make_dev = get_device_fptr(device_handle);