+++ src/utils/utils.hpp
@@ -93,7 +93,7 @@ const std::string default_command_map_name = "command_map";
  * 
  * @note This will have to be manually changed after the release
  */
-const std::string current_host_app_version = "1.0.0";
+const std::string current_host_app_version = "1.1.0";
 
 /** @brief Convert string to uper case */
 std::string to_upper(std::string str);