+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Host XVF Control change log
 0.2.0
 -----
 
+  * FIXED: ``--execute-command-list`` option
   * ADDED: ``hidden_cmd`` flag in ``cmd_t`` structure
   * ADDED: Builds for x86 Mac and Windows
   * CHANGED: UI for ``--help`` and ``--list-commands`` options